Magic Variable and Magic Object Class
"Magic" variable added to stealthCharacter_BP. This will be used for health, time, and abilities.

HUD displays "Magic" as text on lower center of screen.

"Magic" decreases on the tick event in SC_BP.

"MagicObject" class created. This will be used for Actors that hold magic and can be drained. Other functions may be connected to this such as shaking when magic is used near them and such.